Fix bug in dynamic HTTPRouterHandler switching to prevent StackOverflow.
