+++ bifromq-apiserver/src/test/java/com/baidu/bifromq/apiserver/http/HTTPRequestRouterTest.java
@@ -13,39 +13,40 @@
 
 package com.baidu.bifromq.apiserver.http;
 
-import static com.baidu.bifromq.apiserver.Headers.HEADER_REQ_ID;
-import static com.baidu.bifromq.apiserver.Headers.HEADER_TENANT_ID;
 import static io.netty.handler.codec.http.HttpHeaderNames.CONNECTION;
 import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_LENGTH;
 import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_TYPE;
 import static io.netty.handler.codec.http.HttpHeaderValues.CLOSE;
 import static io.netty.handler.codec.http.HttpHeaderValues.KEEP_ALIVE;
+import static io.netty.handler.codec.http.HttpResponseStatus.BAD_REQUEST;
+import static io.netty.handler.codec.http.HttpResponseStatus.OK;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
+import com.baidu.bifromq.apiserver.Headers;
 import com.baidu.bifromq.apiserver.MockableTest;
 import com.baidu.bifromq.plugin.settingprovider.ISettingProvider;
 import com.baidu.bifromq.plugin.settingprovider.Setting;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.embedded.EmbeddedChannel;
 import io.netty.handler.codec.http.DefaultFullHttpRequest;
-import io.netty.handler.codec.http.DefaultHttpRequest;
+import io.netty.handler.codec.http.DefaultFullHttpResponse;
 import io.netty.handler.codec.http.HttpMethod;
 import io.netty.handler.codec.http.HttpResponse;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.netty.handler.codec.http.HttpVersion;
 import java.util.concurrent.CompletableFuture;
+import lombok.extern.slf4j.Slf4j;
 import org.mockito.Mock;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+@Slf4j
 public class HTTPRequestRouterTest extends MockableTest {
     private EmbeddedChannel channel;
     @Mock
@@ -54,6 +55,8 @@ public class HTTPRequestRouterTest extends MockableTest {
     private IHTTPRequestHandler requestHandler;
     @Mock
     private ISettingProvider settingProvider;
+
+    private String tenantId = "tenantId";
 
     @BeforeMethod
     public void setup() {
@@ -65,23 +68,19 @@ public class HTTPRequestRouterTest extends MockableTest {
             }
         });
         channel.freezeTime();
-    }
-
-    @Test
-    public void badRequest() {
-        DefaultHttpRequest req = new DefaultHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
-        channel.writeInbound(req);
-        HttpResponse resp = channel.readOutbound();
-        assertEquals(resp.protocolVersion(), req.protocolVersion());
-        assertEquals(resp.status(), HttpResponseStatus.BAD_REQUEST);
-        assertEquals(resp.headers().get(CONTENT_TYPE), "text/plain; charset=UTF-8");
-        assertEquals(resp.headers().getInt(CONTENT_LENGTH), 0);
+        when(settingProvider.provide(Setting.MaxUserPayloadBytes, tenantId)).thenReturn(256 * 1024);
     }
 
     @Test
     public void keepAliveHeaderForHTTP_1_0() {
-        DefaultHttpRequest req = new DefaultHttpRequest(HttpVersion.HTTP_1_0, HttpMethod.GET, "/fake");
+        DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_0, HttpMethod.GET, "/fake");
         req.headers().set(CONNECTION, KEEP_ALIVE);
+        req.headers().set(Headers.HEADER_TENANT_ID.header, tenantId);
+        when(routeMap.getHandler(any())).thenReturn(requestHandler);
+        when(requestHandler.handle(anyLong(), anyString(), any()))
+            .thenReturn(CompletableFuture.completedFuture(
+                new DefaultFullHttpResponse(HttpVersion.HTTP_1_0, HttpResponseStatus.OK)));
+
         channel.writeInbound(req);
         HttpResponse resp = channel.readOutbound();
         assertEquals(resp.protocolVersion(), req.protocolVersion());
@@ -90,85 +89,95 @@ public class HTTPRequestRouterTest extends MockableTest {
 
     @Test
     public void closeConnectionForHTTP_1_0() {
-        DefaultHttpRequest req = new DefaultHttpRequest(HttpVersion.HTTP_1_0, HttpMethod.GET, "/fake");
+        DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_0, HttpMethod.GET, "/fake");
+        req.headers().set(Headers.HEADER_TENANT_ID.header, tenantId);
+        when(routeMap.getHandler(any())).thenReturn(requestHandler);
+        when(requestHandler.handle(anyLong(), anyString(), any()))
+            .thenReturn(CompletableFuture.completedFuture(
+                new DefaultFullHttpResponse(HttpVersion.HTTP_1_0, HttpResponseStatus.OK)));
+
         channel.writeInbound(req);
         HttpResponse resp = channel.readOutbound();
         assertEquals(resp.protocolVersion(), req.protocolVersion());
+        assertEquals(resp.status().code(), OK.code());
         assertEquals(resp.headers().get(CONNECTION), CLOSE.toString());
         assertFalse(channel.isOpen());
     }
 
     @Test
-    public void closeConnectionByRequest() {
-        DefaultHttpRequest req = new DefaultHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
+    public void handleNormalReturn() {
+        DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
+        long reqId = System.nanoTime();
+        req.headers().set(Headers.HEADER_REQ_ID.header, reqId);
+        req.headers().set(Headers.HEADER_TENANT_ID.header, tenantId);
+        when(routeMap.getHandler(any())).thenReturn(requestHandler);
+        when(requestHandler.handle(anyLong(), anyString(), any()))
+            .thenReturn(CompletableFuture.completedFuture(
+                new DefaultFullHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK)));
+
+        channel.writeInbound(req);
+
+        HttpResponse resp = channel.readOutbound();
+        assertEquals(resp.protocolVersion(), req.protocolVersion());
+        assertEquals(resp.status(), HttpResponseStatus.OK);
+        assertEquals(resp.headers().get(Headers.HEADER_REQ_ID.header), Long.toString(reqId));
+    }
+
+    @Test
+    public void requestWithoutTenantId() {
+        DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
         req.headers().set(CONNECTION, CLOSE);
+        when(routeMap.getHandler(any())).thenReturn(requestHandler);
+        when(requestHandler.handle(anyLong(), anyString(), any()))
+            .thenReturn(CompletableFuture.completedFuture(
+                new DefaultFullHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK)));
+
         channel.writeInbound(req);
         HttpResponse resp = channel.readOutbound();
         assertEquals(resp.protocolVersion(), req.protocolVersion());
         assertEquals(resp.headers().get(CONNECTION), CLOSE.toString());
+        assertEquals(resp.status().code(), BAD_REQUEST.code());
+
         assertFalse(channel.isOpen());
     }
 
     @Test
-    public void setupRouterForTenant() {
-        String tenantId = "bifromq-dev";
+    public void handleException() {
         DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
-        req.headers().set(HEADER_TENANT_ID.header, tenantId);
-        req.headers().set(CONTENT_LENGTH, 0);
-        when(settingProvider.provide(Setting.MaxUserPayloadBytes, tenantId)).thenReturn(
-            Setting.MaxUserPayloadBytes.current(tenantId));
+        long reqId = System.nanoTime();
+        req.headers().set(Headers.HEADER_REQ_ID.header, reqId);
+        req.headers().set(Headers.HEADER_TENANT_ID.header, tenantId);
         when(routeMap.getHandler(any())).thenReturn(requestHandler);
         when(requestHandler.handle(anyLong(), anyString(), any())).thenReturn(
             CompletableFuture.failedFuture(new RuntimeException("Mocked Exception")));
+
+        channel.writeInbound(req);
+
+        HttpResponse resp = channel.readOutbound();
+        assertEquals(resp.protocolVersion(), req.protocolVersion());
+        assertEquals(resp.status(), HttpResponseStatus.INTERNAL_SERVER_ERROR);
+        assertEquals(resp.headers().get(CONTENT_TYPE), "text/plain; charset=UTF-8");
+        assertTrue(resp.headers().getInt(CONTENT_LENGTH) > 0);
+        assertEquals(resp.headers().get(Headers.HEADER_REQ_ID.header), Long.toString(reqId));
+    }
+
+    @Test
+    public void tooLargeRequest() {
+        DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
+        long reqId = System.nanoTime();
+        req.headers().set(Headers.HEADER_REQ_ID.header, reqId);
+        req.headers().set(Headers.HEADER_TENANT_ID.header, tenantId);
+        req.headers().set(CONTENT_LENGTH, 1024 * 2048);
+        when(routeMap.getHandler(any())).thenReturn(requestHandler);
+        when(requestHandler.handle(anyLong(), anyString(), any()))
+            .thenReturn(CompletableFuture.completedFuture(
+                new DefaultFullHttpResponse(HttpVersion.HTTP_1_1, HttpResponseStatus.OK)));
+
         channel.writeInbound(req);
         HttpResponse resp = channel.readOutbound();
         assertEquals(resp.protocolVersion(), req.protocolVersion());
-        assertTrue(resp.headers().contains(HEADER_REQ_ID.header));
-        assertEquals(req.refCnt(), 0);
-    }
+        assertEquals(resp.status(), HttpResponseStatus.REQUEST_ENTITY_TOO_LARGE);
 
-    @Test
-    public void reuseExistingRouter() {
-        String tenantId = "bifromq-dev";
-        DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
-        req.headers().set(HEADER_TENANT_ID.header, tenantId);
-        req.headers().set(CONTENT_LENGTH, 0);
-        when(settingProvider.provide(Setting.MaxUserPayloadBytes, tenantId)).thenReturn(
-            Setting.MaxUserPayloadBytes.current(tenantId));
-        when(routeMap.getHandler(any())).thenReturn(requestHandler);
-        when(requestHandler.handle(anyLong(), anyString(), any())).thenReturn(
-            CompletableFuture.failedFuture(new RuntimeException("Mocked Exception")));
-
-        channel.writeInbound(req);
-
-        req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
-        req.headers().set(HEADER_TENANT_ID.header, tenantId);
-        req.headers().set(CONTENT_LENGTH, 0);
-        channel.writeInbound(req);
-        verify(settingProvider, times(1)).provide(Setting.MaxUserPayloadBytes, tenantId);
-    }
-
-    @Test
-    public void routerSwitch() {
-        String tenantA = "bifromq-dev";
-        String tenantB = "bifromq-ops";
-        DefaultFullHttpRequest req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
-        req.headers().set(HEADER_TENANT_ID.header, tenantA);
-        req.headers().set(CONTENT_LENGTH, 0);
-        when(settingProvider.provide(Setting.MaxUserPayloadBytes, tenantA)).thenReturn(
-            Setting.MaxUserPayloadBytes.current(tenantA));
-        when(settingProvider.provide(Setting.MaxUserPayloadBytes, tenantB)).thenReturn(
-            Setting.MaxUserPayloadBytes.current(tenantB));
-        when(routeMap.getHandler(any())).thenReturn(requestHandler);
-        when(requestHandler.handle(anyLong(), anyString(), any())).thenReturn(
-            CompletableFuture.failedFuture(new RuntimeException("Mocked Exception")));
-
-        channel.writeInbound(req);
-
-        req = new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, HttpMethod.GET, "/fake");
-        req.headers().set(HEADER_TENANT_ID.header, tenantB);
-        req.headers().set(CONTENT_LENGTH, 0);
-        channel.writeInbound(req);
-        verify(settingProvider, times(2)).provide(any(), anyString());
+        assertFalse(channel.isOpen());
     }
 }
